Changing files and classes names so that they follow the PEP8 standard.
Made the package separated into more discrete folders so that it is
possible to better classify what belong where.

Changed name of the package to aiida-uppasd, independently of issues
with the previous version there is no need for a different name.

Fixed several typos.

Added a gitignore to avoid unnecessary files being added to the
repository.

Removed unnecessesary pre-compiled code from the repo.